LDDM 31_10 TP2_CRUD_Links

+++ Extraclasse/android/Divulgar/app/src/main/java/com/example/geovane/divulgar/Model/Curso.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
  */
 
 public class Curso implements Serializable {
-    private int id;
+    private long id;
     private String nomeCurso;
     private ArrayList<Periodo> periodos;
 
@@ -16,17 +16,17 @@ public class Curso implements Serializable {
         this(1, nomeCurso);
     }
 
-    public Curso(int id, String nomeCurso) {
+    public Curso(long id, String nomeCurso) {
         this.id = id;
         this.nomeCurso = nomeCurso;
         this.periodos = new ArrayList<Periodo>();
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
